# specs/2025-08-14_OPENAPI.yaml
openapi: 3.1.0
info:
  title: GeoGuide API
  version: "1.0.0"
  description: >
    API per guida geolocalizzata multilingua (LangGraph + MongoDB).
    Nota: la collection `narrations_cache` è persistente con **TTL configurabile** (default 24h).
servers:
  - url: https://api.geoguide.example.com/v1
...
paths:
...
  /narration:
    post:
      tags: [Narrations]
      summary: Genera o recupera narrazione
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                poi_id: { type: string }
                lang: { type: string }
                style: { type: string, enum: [guide, quick, kids, anecdotes] }
              required: [poi_id, lang, style]
      responses:
        '200':
          description: Narrazione
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Narration' }
    get:
      tags: [Narrations]
      summary: Recupera narrazione da cache
      parameters:
        - in: query
          name: poi_id
          schema: { type: string }
          required: true
        - in: query
          name: lang
          schema: { type: string }
          required: true
        - in: query
          name: style
          schema: { type: string }
          required: true
      responses:
        '200':
          description: Narrazione
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Narration' }
...
components:
  schemas:
...
    Narration:
      type: object
      properties:
        poi_id: { type: string }
        style: { type: string }
        lang: { type: string }
        text: { type: string }
        audio_url: { type: string }
        sources:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              url: { type: string }
        confidence: { type: number }
        created_at: { type: string, format: date-time }
...